requirements.txt
streamlit
numpy
opencv-python
pillow
tensorflow
ultralytics
opencv-python-headless

